==> scripts/13_Instrumental_Clinical_Cis.ipynb <==
# Instrumental analysis requiring instruments in *cis*

## Libraries

library(readr)
library(dplyr, warn.conflicts = FALSE)
library(tidyr, warn.conflicts = FALSE)
library(purrr)
library(ggplot2)

## Reference panel

freq_1kg <- read_tsv("~/projects/DVA/Data/ReferenceData/1kg.afreq",
                     skip = 1,
                     col_names = c("chr", "position", "rsid", "nea_1kg_plus", "ea_1kg_plus", "eaf_1kg"),
                     col_types = "nncccn")

head(freq_1kg)

## Outcome data - T2D

outcome_data <- read_tsv(
    pipe("sed 's/:/\t/g' ~/projects/DVA/Data/GWAS_sumstats/t2d_diagram.txt"),
    skip = 1,
    col_names = c("chr", "position", 
                  paste(c("ea", "nea", "beta", "se", "p", "n"), "t2d", sep = "_")),
    col_types = "nnccnnnn"
)
head(outcome_data)

Harmonizing outcome data to reference panel:

outcome_data <- outcome_data %>%
    inner_join(freq_1kg, by = c("chr", "position")) %>%
    mutate(
        ea_1kg_minus = stringr::str_replace_all(ea_1kg_plus, c("A" = "t", "T" = "a", "C" = "g", "G" = "c")),
        ea_1kg_minus = toupper(ea_1kg_minus),
        nea_1kg_minus = stringr::str_replace_all(nea_1kg_plus, c("A" = "t", "T" = "a", "C" = "g", "G" = "c")),
        nea_1kg_minus = toupper(nea_1kg_minus),
        harmon = case_when(ea_t2d == ea_1kg_plus & nea_t2d == nea_1kg_plus ~ 1,
                           ea_t2d == ea_1kg_minus & nea_t2d == nea_1kg_minus ~ 1,
                           ea_t2d == nea_1kg_plus & nea_t2d == ea_1kg_plus ~ -1,
                           ea_t2d == nea_1kg_minus & nea_t2d == ea_1kg_minus ~ -1,
                           TRUE ~ 0)
    ) %>%
    filter(harmon != 0) %>%
    select(-c(harmon, ea_1kg_plus, ea_1kg_minus, nea_1kg_plus, nea_1kg_minus)) %>%
    # Ambiguous palindromic SNPs
    mutate(maf_1kg = ifelse(eaf_1kg < .5, eaf_1kg, 1 - eaf_1kg),
           palind_1 = ea_t2d %in% c("A", "T") & nea_t2d %in% c("A", "T"),
           palind_2 = ea_t2d %in% c("C", "G") & nea_t2d %in% c("C", "G"),
           palind_ambig = (palind_1 | palind_2) & maf_1kg > .4) %>%
    filter(!palind_ambig) %>%
    select(-starts_with("palind"), -eaf_1kg)
head(outcome_data)

## Exposure data

### BMI

bmi <- read_tsv(
    "~/projects/DVA/Data/GWAS_sumstats/bmi.txt", 
    skip = 1,
    col_names = c("chr", "position", "rsid", 
                  paste(c("ea", "nea", "eaf", "beta", "se", "pval", "n"), "bmi", sep = "_")),
    col_types = "nncccnnnnn"
)
head(bmi)

### Data for each trait

Exposure data:

sigins <- read_tsv(
    "../data/sigins.tsv", 
    col_types = "nnnnnnccccnc"
)
head(sigins)

To select instruments:

- Gene IDs:

exposures_cismr <- tibble::tribble(
    ~trait_short, ~requireCisMR, ~id, ~ensembl_id,
    "SHBG", TRUE, "ukb-d-30830_irnt", "ENSG00000129214",
    "GGT", TRUE, "ukb-d-30730_irnt", paste(c("ENSG00000149435",
                                             "ENSG00000100121", 
                                             "ENSG00000274252", 
                                             "ENSG00000230712", 
                                             "ENSG00000276160", 
                                             "ENSG00000100031", 
                                             "ENSG00000133475", 
                                             "ENSG00000197421", 
                                             "ENSG00000280208", 
                                             "ENSG00000099998", 
                                             "ENSG00000167741", 
                                             "ENSG00000131067", 
                                             "ENSG00000236969"),
                                           collapse = ";"),
    "ALT", TRUE, "ukb-d-30620_irnt", paste(c("ENSG00000167701", "ENSG00000166123"), collapse = ";"),
    "ApoA1", TRUE, "ieu-b-107", "ENSG00000118137",
    "PTH", TRUE, "prot-a-2431", "ENSG00000152266",
    "CRP", TRUE, "ukb-d-30710_irnt", "ENSG00000132693"
)

exposures_cismr <- separate_rows(exposures_cismr, ensembl_id)

head(exposures_cismr)

- Genomic regions in hg19 build:

gencode <- rtracklayer::readGFF("~/projects/DVA/Data/ReferenceData/gencode.v19.annotation.gtf")
head(gencode)

- Joining exposure IDs with genomic regions:

exposures <- gencode %>%
    filter(type == "gene") %>%
    transmute(chrom = as.numeric(gsub("chr", "", seqid)), start, end, ensembl_id = gsub("\\..*", "", gene_id)) %>%
    inner_join(exposures_cismr, by = "ensembl_id")
head(exposures)

- Now selecting instruments:

sigins_cis <- sigins %>%
    inner_join(exposures) %>%
    filter(chrom == chr) %>%
    mutate(within_gene = position > start & position < end,
           dist1 = abs(position - start),
           dist2 = abs(position - end)) %>%
    filter(within_gene | dist1 < 5e5 | dist2 < 5e5) %>%
    inner_join(bmi) %>%
    filter(pval_bmi < 5e-8)
head(sigins_cis)

sigins_cis$trait_short %>% unique

- Selecting instruments with strong effects on BMI and harmonizing:

validins_bmi <- sigins_cis %>%
    rename_with(~paste0(.x, "_exp"), c(ea, nea, eaf, beta, se, p)) %>%
    inner_join(outcome_data) %>%
    mutate(maf_bmi = ifelse(eaf_bmi < .5, eaf_bmi, 1 - eaf_bmi),
           new_ea = ifelse(beta_bmi > 0, ea_bmi, nea_bmi),
           new_nea = ifelse(beta_bmi > 0, nea_bmi, ea_bmi),
           eaf_bmi = ifelse(beta_bmi > 0, eaf_bmi, 1 - eaf_bmi),
           beta_bmi = abs(beta_bmi)) %>%
    select(-c(ea_bmi, nea_bmi)) %>%
    filter(abs(maf_bmi - maf_1kg) < .2) %>%
    ## Aligning to the BMI increasing allele
    mutate(
        ea_flip = toupper(stringr::str_replace_all(new_ea, c("A" = "t", "T" = "a", "C" = "g", "G" = "c"))),
        nea_flip = toupper(stringr::str_replace_all(new_nea, c("A" = "t", "T" = "a", "C" = "g", "G" = "c"))),
        harmon = case_when(ea_exp == new_ea & nea_exp == new_nea ~ 1,
                           ea_exp == new_nea & nea_exp == new_ea ~ -1,
                           ea_flip == new_ea & nea_flip == new_nea ~ 1,
                           ea_flip == new_nea & nea_flip == new_ea ~ -1,
                           TRUE ~ 0),
        beta_exp = beta_exp * harmon,
        beta_t2d = beta_t2d * harmon,
        ea = new_ea, nea = new_nea
    ) %>%
    ## Removing SNPs with allele mismatch
    filter(harmon != 0, abs(maf_bmi - maf_1kg) < .2) %>%
    select(-c(harmon, new_ea, new_nea, ea_flip, nea_flip))
head(validins_bmi)

validins_bmi %>%
    

- Clumping to obtain independent SNPs:

clump_fx <- function(rsid, pval, 
                     plink = "/ludc/Tools/Software/Plink/v1.90b5.2/plink", 
                     bfile = "/ludc/Home/daniel_c/projects/DVA/Data/ReferenceData/1kg_ref/EUR",
                     clump_p = 5e-8, clump_kb = 500, clump_r2 = 0.01){
    fn <- tempfile()
    dat <- data.frame(SNP = rsid, P = pval)
    write_tsv(dat, fn)
    plink_cmd <- paste0(plink,
                        " --bfile ", bfile,
                        " --clump ", fn, 
                        " --clump-p1 ", clump_p, 
                        " --clump-r2 ", clump_r2,
                        " --clump-kb ", clump_kb, 
                        " --out ", fn)
    system(plink_cmd)
    clumped_snps <- paste("awk '{print $3}'", paste(fn, "clumped", sep = ".")) %>%
        pipe %>%
        readLines
    unlink(paste0(fn, "*"))
    return(clumped_snps)
}

cis_ins_dat <- validins_bmi %>%
    group_by(trait_short) %>%
    group_modify(~{
        snps_to_retain <- clump_fx(.x$rsid, .x$p_exp)
        clumped_dat <- filter(.x, rsid %in% snps_to_retain)
    }) %>%
    ungroup
cis_ins_dat

write_tsv(cis_ins_dat, "../data/cis_ins_dat.tsv")

cis_ins_dat %>%
    data.frame %>%
    print
